pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asgi-correlation-id>=4.3.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "pyyaml>=6.0.3",
    "result>=0.17.0",
    "sqlalchemy>=2.0.48",
    "tenacity>=9.1.4",
    "uuid6>=2025.0.1",
    "uvicorn>=0.41.0",
]
